client-side/requirements.txt
PyQt5
psutil
rsa
requests[socks]
pycryptodome
